refactor(engine): decompose ecs/handlers into packages and slim the Range model
Finish the engine god-module decomposition (#685). The engine.services surface
was already a package; this addresses the remaining concentration:

- engine/ecs.py -> engine/ecs/ package (facade __init__ + _env, _config,
  _local) split by responsibility. Dispatch continues through the existing
  shared.cloud TaskRunner port (coordinates with the closed #498); the stable
  "engine.ecs" logger namespace and patch-at-source seams are preserved.
- engine/handlers.py -> engine/handlers/ package (routing facade + _range,
  _ngfw, _audit). Preserves the engine.handlers.process_event dotted path,
  the parse_sns_message re-export, audit source literals, and ADR-025 retry
  semantics (transient failures raise; permanent/unknown messages log-return).
- Move the Range model's provisioned_instances projection and status
  classification into the dependency-neutral engine/_range_state.py
  (re-exported by services._common/_lifecycle), removing raw traversal and a
  duplicate status vocabulary from the ORM model while keeping ORM query
  classmethods and the subnet-index concurrency helper.

Behaviour-preserving: no public API, wire, persistence, auth, or provider
change. Adds pure-function unit tests plus a local-provisioner default-path
regression test; existing engine and cross-domain suites stay green.
